#31 Abstract imageviewer og simpel imageviewer
