+++ app/src/main/AndroidManifest.xml
@@ -17,15 +17,21 @@
         <activity android:name=".main.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
+
                 <action android:name="android.intent.action.SEARCH" />
             </intent-filter>
-            <meta-data android:name="android.app.searchable"
+
+            <meta-data
+                android:name="android.app.searchable"
                 android:resource="@xml/search_config" />
         </activity>
         <activity android:name=".item.ItemActivity" />
         <activity android:name=".image.ImageviewerDeleteActivity" />
         <activity android:name=".item.RecordingActivity" />
+        <activity android:name=".image.ImageviewerSimpleActivity" />
+        <activity android:name=".image.AbstractImageViewer"></activity>
     </application>
 
 </manifest>
